fix(frontend): apply minor fixes from review
