+++ frontend_lib/src/action.async.js
@@ -1,4 +1,5 @@
-import { COMMON_REQUEST_HEADERS, CONTENT_NAMESPACE, FETCH_CONFIG, CONTENT_TYPE } from './helper.js'
+import { COMMON_REQUEST_HEADERS, FETCH_CONFIG } from './helper.js'
+import { CONTENT_NAMESPACE, CONTENT_TYPE } from './helperConstants.js'
 
 export const baseFetch = (method, url, body = undefined) => {
   const isFormData = body instanceof FormData
