+++ frontend_lib/src/index.js
@@ -1,31 +1,5 @@
 import {
-  ACCESSIBLE_SPACE_TYPE_LIST,
-  ALLOWED_CHARACTERS_USERNAME,
-  APP_FEATURE_MODE,
-  BREADCRUMBS_TYPE,
-  CHECK_USERNAME_DEBOUNCE_WAIT,
-  COLLABORA_EXTENSIONS,
-  CONTENT_NAMESPACE,
-  CONTENT_TYPE,
-  DATE_FNS_LOCALE,
   FETCH_CONFIG,
-  FILE_PREVIEW_STATE,
-  IMG_LOAD_STATE,
-  MAXIMUM_CHARACTERS_USERNAME,
-  MINIMUM_CHARACTERS_USERNAME,
-  NUMBER_RESULTS_BY_PAGE,
-  PAGE,
-  PROFILE_LIST,
-  PROFILE,
-  ROLE_LIST,
-  ROLE,
-  SPACE_TYPE_LIST,
-  SPACE_TYPE,
-  STATUSES,
-  SUBSCRIPTION_TYPE_LIST,
-  SUBSCRIPTION_TYPE,
-  TIMELINE_TYPE,
-  USER_CALL_STATE,
   addAllResourceI18n,
   addExternalLinksIcons,
   addRevisionFromTLM,
@@ -61,7 +35,6 @@ import {
   permissiveNumberEqual,
   removeAtInUsername,
   removeExtensionOfFilename,
-  revisionTypeList,
   scrollIntoViewIfNeeded,
   sendGlobalFlashMessage,
   serialize,
@@ -73,6 +46,37 @@ import {
   stripEmojis,
   updateTLMUser
 } from './helper.js'
+
+import {
+  ACCESSIBLE_SPACE_TYPE_LIST,
+  ALLOWED_CHARACTERS_USERNAME,
+  APP_FEATURE_MODE,
+  BREADCRUMBS_TYPE,
+  CHECK_USERNAME_DEBOUNCE_WAIT,
+  COLLABORA_EXTENSIONS,
+  CONTENT_NAMESPACE,
+  CONTENT_TYPE,
+  DATE_FNS_LOCALE,
+  FILE_PREVIEW_STATE,
+  IMG_LOAD_STATE,
+  MAXIMUM_CHARACTERS_USERNAME,
+  MINIMUM_CHARACTERS_USERNAME,
+  NUMBER_RESULTS_BY_PAGE,
+  PAGE,
+  PROFILE_LIST,
+  PROFILE,
+  revisionTypeList,
+  ROLE_LIST,
+  ROLE,
+  SPACE_TYPE_LIST,
+  SPACE_TYPE,
+  STATUSES,
+  SUBSCRIPTION_TYPE_LIST,
+  SUBSCRIPTION_TYPE,
+  TIMELINE_TYPE,
+  USER_CALL_STATE
+} from './helperConstants.js'
+
 import i18n, { getBrowserLang } from './i18n.js'
 
 import {
