+++ frontend_lib/test/component/Timeline.spec.js
@@ -4,7 +4,7 @@ import { expect } from 'chai'
 import { mount } from 'enzyme'
 import { Timeline } from '../../src/component/Timeline/Timeline.jsx'
 import sinon from 'sinon'
-import { ROLE } from '../../src/helper.js'
+import { ROLE } from '../../src/helperConstants.js'
 import { commentList } from '../fixture/contentCommentList.js'
 import { revisionList } from '../fixture/contentRevisionList.js'
 import { reactstrapPopoverHack } from '../testHelper.js'
