+++ frontend_lib/test/sortListHelper.spec.js
@@ -1,5 +1,5 @@
 import { expect } from 'chai'
-import { CONTENT_TYPE } from '../src/helper.js'
+import { CONTENT_TYPE } from '../src/helperConstants.js'
 import {
   sortWithFoldersAtListBeginning,
   SORT_BY,
